+doc
git-svn-id: svn://svn.code.sf.net/p/genomeview/code/genomeview/trunk@1332 6512a8d4-c03c-0410-a19d-ba0f7ff4f5ce

+++ net/sf/genomeview/gui/annotation/track/PileupTrack.java
@@ -190,8 +190,9 @@ public class PileupTrack extends Track {
 	 * value will be reset each time we retrieve new data.
 	 */
 	private double localMaxPile = 0;
-
-	private int maxSummary = 0;
+	/* Keeps track of the maximum value in summary graph mode */
+	private double maxSummary = 0;
+	
 
 	private static final int CHUNK = 32000;
 	private static final int SUMMARYSIZE = 100;
